Link with PsyQ SDK

+++ src/memcard/memcard.h
@@ -23,7 +23,7 @@ typedef void (*TMemCardFunc)(int);
 
 void memcard_init_80024E48();
 void memcard_exit_800250C4();
-int  memcard_check_80024A54(int idx);
+long memcard_check_80024A54(long idx);
 int  memcard_easy_format_test_800246C0(int hCard);
 void memcard_reset_status_80024A3C();
 void memcard_set_sw_hw_card_fns_8002469C();
